@@ -17,7 +17,7 @@
         }
     },
     "require": {
-        "php": "^7.3||~8.0||~8.1",
+        "php": "^7.4||~8.0||~8.1",
         "wildbit/postmark-php": "^2.9||^3.0||^4.0",
         "psr/container": "^1.0.0 || ^2.0.0"
     },
@@ -33,8 +33,8 @@
     },
     "require-dev": {
         "doctrine/coding-standard": "^9.0",
-        "laminas/laminas-component-installer": "^2.5.1",
-        "laminas/laminas-servicemanager": "^3.7.0",
+        "laminas/laminas-component-installer": "^2.8.0",
+        "laminas/laminas-servicemanager": "^3.16.0",
         "lctrs/psalm-psr-container-plugin": "^1.6",
         "phpunit/phpunit": "^9.5.23",
         "psalm/plugin-phpunit": "^0.17.0",
@@ -44,12 +44,14 @@
     "scripts": {
         "check": [
             "@cs-check",
+            "psalm --stats",
             "@test"
         ],
         "cs-check": "phpcs",
         "cs-fix": "phpcbf",
         "test": "phpunit --colors=always",
-        "test-coverage": "phpunit --colors=always --coverage-html build/report"
+        "test-coverage": "phpunit --colors=always --coverage-html build/report",
+        "static-analysis": "psalm --stats --shepherd"
     },
     "extra": {
         "laminas": {
